refactor: Http Response 패키지 이동 - http 패키지의 위치를 global로 변경했어요
