+++ src/main/java/com/group2/filterism/domain/hashtag/presentation/HashtagReadController.java
@@ -1,7 +1,7 @@
 package com.group2.filterism.domain.hashtag.presentation;
 
 import com.group2.filterism.domain.hashtag.application.HashtagReadUseCase;
-import com.group2.filterism.http.ListResponse;
+import com.group2.filterism.global.http.ListResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
